# go/docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  models.Kermesse:
    properties:
      id:
        type: integer
      name:
        type: string
      organisateurs:
        description: 'Relations Many-to-Many : Organisateurs et participants de la
          kermesse'
        items:
          $ref: '#/definitions/models.User'
        type: array
      participants:
        items:
          $ref: '#/definitions/models.User'
        type: array
      stands:
        items:
          $ref: '#/definitions/models.Stand'
        type: array
      user:
        $ref: '#/definitions/models.User'
      user_id:
        description: 'Relation Many-to-One : L''utilisateur qui crée la kermesse'
        type: integer
    type: object
  models.Product:
    properties:
      id:
        type: integer
      jetons_requis:
        type: integer
      name:
        type: string
      nb_products:
        type: integer
      picture:
        type: string
      stand_id:
        description: Clé étrangère vers le stand
        type: integer
      type:
        type: string
    type: object
  models.Stand:
    properties:
      conso:
        type: integer
      id:
        type: integer
      jetons_requis:
        type: integer
      kermesses:
        items:
          $ref: '#/definitions/models.Kermesse'
        type: array
      name:
        type: string
      pts_donnees:
        type: integer
      stocks:
        description: Clé étrangère vers Product
        items:
          $ref: '#/definitions/models.Product'
        type: array
      type:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.Transaction:
    properties:
      Quantity:
        type: integer
      date_transaction:
        type: string
      id:
        type: integer
      price:
        type: number
      type:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Association avec l'utilisateur
      user_id:
        description: Relations avec l'utilisateur
        type: integer
    type: object
  models.User:
    properties:
      email:
        type: string
      enfants:
        items:
          $ref: '#/definitions/models.User'
        type: array
      firstname:
        type: string
      id:
        type: integer
      jetons:
        type: integer
      kermesses:
        description: Relations One-to-Many pour les Kermesses et Stands gérés par
          l'utilisateur
        items:
          $ref: '#/definitions/models.Kermesse'
        type: array
      lastname:
        type: string
      parents:
        description: Relations Many-to-Many pour Parents/Enfants
        items:
          $ref: '#/definitions/models.User'
        type: array
      password:
        type: string
      picture:
        type: string
      role:
        description: 1 = ADMIN / 2 = ORGANISATEUR / 3 = TENEUR DE STAND / 4 = PARENT
          / 5 ELEVE
        type: integer
      stands:
        items:
          $ref: '#/definitions/models.Stand'
        type: array
      transactions:
        items:
          $ref: '#/definitions/models.Transaction'
        type: array
    type: object
...
  requests.SignupRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      picture:
        type: string
      role:
        type: integer
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
...
info:
  contact: {}
paths:
...
  /api/users:
    get:
      description: Récupère la liste de tous les utilisateurs
      parameters:
      - default: Bearer Add access token here
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Erreur serveur interne
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Récupère tous les utilisateurs
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Crée un nouvel utilisateur avec les informations fournies
      parameters:
      - default: Bearer Add access token here
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Utilisateur à créer
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/requests.SignupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "500":
          description: Erreur serveur interne
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Crée un nouvel utilisateur
      tags:
      - User
  /api/users/{id}:
    delete:
      description: Supprime un utilisateur spécifique
      parameters:
      - default: Bearer Add access token here
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID de l'utilisateur
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Utilisateur non trouvé
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Erreur serveur interne
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Supprime un utilisateur par ID
      tags:
      - User
    get:
      description: Récupère les informations d'un utilisateur spécifique
      parameters:
      - default: Bearer Add access token here
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID de l'utilisateur
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Utilisateur non trouvé
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Erreur serveur interne
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Récupère un utilisateur par ID
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Met à jour les informations d'un utilisateur spécifique
      parameters:
      - default: Bearer Add access token here
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID de l'utilisateur
        in: path
        name: id
        required: true
        type: integer
      - description: Utilisateur à mettre à jour
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Utilisateur non trouvé
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Erreur serveur interne
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Met à jour un utilisateur par ID
      tags:
      - User
...
swagger: "2.0"